cmplog runtime for LLVM
