+++ include/sharedmem.h
@@ -31,6 +31,7 @@ void setup_shm(unsigned char dumb_mode);
 void remove_shm(void);
 
 extern int cmplog_mode;
+extern struct cmp_map* cmp_map;
 
 #endif
