+++ src/afl-fuzz-globals.c
@@ -251,6 +251,9 @@ u32                a_extras_cnt;        /* Total number of tokens available */
 
 u8 *(*post_handler)(u8 *buf, u32 *len);
 
+u8* cmplog_binary;
+s32 cmplog_forksrv_pid;
+
 /* hooks for the custom mutator function */
 size_t (*custom_mutator)(u8 *data, size_t size, u8 *mutated_out,
                          size_t max_size, unsigned int seed);
